Fix / simplify roughness-prefilter.
This tool was producing bad output for the monkey roughness map because
the monkey map has an (unnecessary) alpha channel, and the tool was
assuming that the source image has 3 channels.

This fixes that particular case and changes the tool so that it
operates on single-channel data instead of 3-channel data.