Fix a default material instance sampler initialization
This fixes #887. this bug existed since forever, when setting the
samplers on a new material instance from the default material
instance, we were losing the dirty bits (i.e. they were copied),
and so the newly initialized buffer would never be comited.


This adds a setSamplers() methods which does exactly that, and for
symmetry (and more efficiently) we also add setUniforms() to 
UniformBuffer.