disable "reverse resolve" and add some asserts
"reverse resolve" is needed on desktop to emulate
EXT_multisampled_render_to_texture. However, it is not easily supported
on Metal or GLES. So for now we avoid relying on this.

This means that filament requirements are lower than what
EXT_multisampled_render_to_texture guarantees.

Essentially filament doesn't rely on being able to preserve the 
content of a non-ms texture attached to a ms RenderTarget.

Also don't rely on being able to resolve+resize/move, which is
not supported in GLES.